+++ dapp/pages/index.js
@@ -8,7 +8,7 @@ import Ogv from 'components/Ogv'
 import Seo from 'components/strapi/seo'
 import Head from 'next/head'
 import { fetchAPI } from '../lib/api'
-import formatSeo from '../src/utils/seo'
+import formatSeo from '../src/utils/formatSeo'
 import transformLinks from '../src/utils/transformLinks'
 import { Typography, Header, Button } from '@originprotocol/origin-storybook'
 import { assetRootPath } from 'utils/image'
